This fix is part of some work I have been doing to solve multiple issues with an ATI GPU.
This is a sort of a crazy solution, but it seems to work perfectly.

My problem was that all the buffers (vertex, index, texture) kept getting corrupted on win32 using ATI and latest drivers, the solution I found is to set OGRE_GL_MAP_BUFFER_THRESHOLD to be max int (sort of the opposite to the hack you have for the non win32 platforms). This solves all the issues and now everything works. This problems is related to the BillboardSet class somehow (if I disable all the BillboardSet I don't have a problem), I tracked the connection to the fact that BillboardSet only updates and uses a sub part of the vertex\index buffers. From here I found the remarks above OGRE_GL_MAP_BUFFER_THRESHOLD. First I tried to set it to 0 and it didn't solve the issue, then I tried the max int as its value - and EVERYTHING worked.

In my code I added a data member for OGRE_GL_MAP_BUFFER_THRESHOLD named "mMapBufferThreshold" to GLHardwareBufferManager, in its init if we are using win32 and ATI I set to max int, if we are not win32 I set to 0, else to the value of OGRE_GL_MAP_BUFFER_THRESHOLD (default).

--HG--
branch : v1-6