SampleBrowser: drop OIS dependancy and use SDL2 on desktop
introduce a thin abstraction over SDL2 that is just SDL2 on desktop and
much simpler to use than OIS on platforms where we just inject events.

the RenderWindow is now also created with SDL so it can handle input
events.

Touch events are redirect as mouse events instead of ifdeffing to allow
using mouse and touch events at the same time.